✨ Update EmployerLocationOption type in Location model
- Changed EmployerLocationOption from short to AvailableWhere.
- Updated UpsertApplicationCommandHandler to cast command value.
- Adjusted handling of employer location options for better clarity.
- Ensured compatibility with existing command structure.
- Improved data representation for employment location options.

Changes made by Balaji Jambulingam

+++ src/SFA.DAS.CandidateAccount.Application/Application/Commands/UpsertApplication/UpsertApplicationCommandHandler.cs
@@ -126,7 +126,7 @@ public class UpsertApplicationCommandHandler(
             Id = Guid.NewGuid(),
             ApplicationId = application.Id,
             Addresses = command.EmploymentLocation.Addresses,
-            EmployerLocationOption = command.EmploymentLocation.EmployerLocationOption,
+            EmployerLocationOption = (short)command.EmploymentLocation.EmployerLocationOption,
             EmploymentLocationInformation = command.EmploymentLocation.EmploymentLocationInformation
         }, command.CandidateId, cancellationToken);
     }
